// examples/typescript/src/generated/inputs/create-person-input.ts
// GENERATED CODE - DO NOT MODIFY BY HAND. ANY MODIFICATION WILL BE LOST ON NEXT GENERATION
// Generated by GraphLink
// GitHub: https://github.com/Oualitsen/graphlink
// Site: https://graphlink.dev
// Pub.dev https://pub.dev/packages/graphlink
// ignore_for_file: use_rethrow_when_possible camel_case_types, constant_identifier_names, unused_import, non_constant_identifier_names



export interface CreatePersonInput {
   name: string;
   age: number;
}